+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Refactored Protocol class name to ProtocolVersion because it shadowed typing.Protocol
 - Implemented improved firmware update detection and subsequent restart policy
 - Implemented stale state solutions for Derived Sensors
+- Simplified Metrics sensors by removing the need for _update_internal_state to be overridden in most cases
 - Upgraded `pydantic-settings` from 2.14.2 to 2.15.0
 - Upgraded `pymodbus` from 3.14.0 to 3.15.0